+++ Example/Analytics/DemoAppAnalytics.swift
@@ -28,7 +28,7 @@ internal enum DemoAppAnalyticsParameter: String, CaseIterable {
 public class DemoAppAnalytics {
     public static let notificationName = Notification.Name("com.rakuten.esd.sdk.events.custom")
     open class var sdkVersion: String? {
-        Bundle.miniAppBundle?.infoDictionary?["CFBundleShortVersionString"] as? String
+        Bundle.miniAppSDKBundle.infoDictionary?["CFBundleShortVersionString"] as? String
     }
     internal static let defaultRATAcc = MAAnalyticsConfig(acc: String(Bundle.main.infoDictionary?["RATAccountIdentifier"] as? Int ?? 0) ,
                                                           aid: String(Bundle.main.infoDictionary?["RATAppIdentifier"] as? Int ?? 0) )