+++ Example/Controllers/Settings/ContactsListSettingsTableViewController.swift
@@ -193,6 +193,7 @@ class ContactsListSettingsTableViewController: RATTableViewController {
         }
     }
 
+    // swiftlint:disable function_body_length
     func validateAllValues(index: Int, contactId: String?, textField: UITextField?, name: String, email: String, isNewContact: Bool? = false) {
         if let contactIdTextField = textField {
             if contactIdTextField.isTextFieldEmpty() {
@@ -201,8 +202,26 @@ class ContactsListSettingsTableViewController: RATTableViewController {
                                  contactId: contactIdTextField.text,
                                  contactName: name,
                                  contactEmail: email, isNewContact: isNewContact)
+            } else if name.isValueEmpty() && email.isValueEmpty() {
+                self.editContact(title: "Name and Email ID are empty, please try again",
+                                 index: index,
+                                 contactId: contactIdTextField.text,
+                                 contactName: name,
+                                 contactEmail: email, isNewContact: isNewContact)
+            } else if name.isValueEmpty() && (email.isValueEmpty() || !email.isValidEmail()) {
+                self.editContact(title: "Name is empty and Email ID is invalid, please try again",
+                                 index: index,
+                                 contactId: contactIdTextField.text,
+                                 contactName: name,
+                                 contactEmail: email, isNewContact: isNewContact)
             } else if name.isValueEmpty() {
                 self.editContact(title: "Invalid Name, please try again",
+                                 index: index,
+                                 contactId: contactIdTextField.text,
+                                 contactName: name,
+                                 contactEmail: email, isNewContact: isNewContact)
+            } else if email.isValueEmpty() {
+                self.editContact(title: "Email ID is empty, please try again",
                                  index: index,
                                  contactId: contactIdTextField.text,
                                  contactName: name,