+++ MiniAppCarthage/MiniApp/MiniApp.h
@@ -1,0 +1,22 @@
+#ifdef __OBJC__
+#import <UIKit/UIKit.h>
+#else
+#ifndef FOUNDATION_EXPORT
+#if defined(__cplusplus)
+#define FOUNDATION_EXPORT extern "C"
+#else
+#define FOUNDATION_EXPORT extern
+#endif
+#endif
+#endif
+#import <Foundation/Foundation.h>
+
+//! Project version number for MiniApp.
+FOUNDATION_EXPORT double MiniAppVersionNumber;
+
+//! Project version string for MiniApp.
+FOUNDATION_EXPORT const unsigned char MiniAppVersionString[];
+
+// In this header, you should import all the public headers of your framework using statements like #import <MiniApp/PublicHeader.h>
+
+